+++ src/directives.js
@@ -137,15 +137,21 @@ angular.module('oi.select')
                     }
                 });
 
+                function valueChangedManually() { //case: clean model; prompt + editItem: 'correct'; initial value = defined/undefined
+                    if (editItemIsCorrected) {
+                        element.removeClass('cleanMode');
+                    }
+                    editItemIsCorrected = false;
+                }
+
                 scope.$parent.$watch(attrs.ngModel, function(value, oldValue) {
                     var output = value instanceof Array ? value : value ? [value]: [],
                         promise = $q.when(output);
 
                     modifyPlaceholder();
 
-                    if (value !== oldValue) {
-                        editItemIsCorrected = false;
-                        element.removeClass('cleanMode');
+                    if (oldValue && value !== oldValue) {
+                        valueChangedManually();
                     }
 
                     if (!multiple) {
@@ -250,6 +256,8 @@ angular.module('oi.select')
                     if (!multiple && !options.closeList) {
                         resetMatches({query: true});
                     }
+
+                    valueChangedManually();
 
                     scope.oldQuery = scope.oldQuery || scope.query;
                     scope.query = '';
@@ -577,7 +585,7 @@ angular.module('oi.select')
                             .finally(function(){
                                 scope.showLoader = false;
 
-                                if (options.closeList) {
+                                if (options.closeList && !options.cleanModel) { //case: prompt
                                     $timeout(function() {
                                         setOption(listElement, 0);
                                     });